+++ src/fidesops/ops/schemas/email/email.py
@@ -24,6 +24,8 @@ class EmailActionType(str, Enum):
     PRIVACY_REQUEST_RECEIPT = "privacy_request_receipt"
     PRIVACY_REQUEST_COMPLETE_ACCESS = "privacy_request_complete_access"
     PRIVACY_REQUEST_COMPLETE_DELETION = "privacy_request_complete_deletion"
+    PRIVACY_REQUEST_REVIEW_DENY = "privacy_request_review_deny"
+    PRIVACY_REQUEST_REVIEW_APPROVE = "privacy_request_review_approve"
 
 
 class EmailTemplateBodyParams(Enum):
@@ -69,6 +71,12 @@ class AccessRequestCompleteBodyParams(BaseModel):
     """Body params required for privacy request completion access email template"""
 
     download_links: List[str]
+
+
+class RequestReviewDenyBodyParams(BaseModel):
+    """Body params required for privacy request review deny email template"""
+
+    rejection_reason: Optional[str]
 
 
 class EmailForActionType(BaseModel):