+++ src/fidesops/ops/service/email/email_dispatch_service.py
@@ -17,6 +17,7 @@ from fidesops.ops.schemas.email.email import (
     EmailServiceType,
     FidesopsEmail,
     RequestReceiptBodyParams,
+    RequestReviewDenyBodyParams,
     SubjectIdentityVerificationBodyParams,
 )
 from fidesops.ops.tasks import DatabaseTask, celery_app
@@ -53,6 +54,7 @@ def dispatch_email(
             AccessRequestCompleteBodyParams,
             SubjectIdentityVerificationBodyParams,
             RequestReceiptBodyParams,
+            RequestReviewDenyBodyParams,
             List[CheckpointActionRequired],
         ]
     ] = None,
@@ -85,7 +87,7 @@ def dispatch_email(
     )
 
 
-def _build_email(
+def _build_email(  # pylint: disable=too-many-return-statements
     action_type: EmailActionType,
     body_params: Any,
 ) -> EmailForActionType:
@@ -129,6 +131,20 @@ def _build_email(
         return EmailForActionType(
             subject="Your data has been deleted",
             body=base_template.render(),
+        )
+    if action_type == EmailActionType.PRIVACY_REQUEST_REVIEW_APPROVE:
+        base_template = get_email_template(action_type)
+        return EmailForActionType(
+            subject="Your request has been approved",
+            body=base_template.render(),
+        )
+    if action_type == EmailActionType.PRIVACY_REQUEST_REVIEW_DENY:
+        base_template = get_email_template(action_type)
+        return EmailForActionType(
+            subject="Your request has been denied",
+            body=base_template.render(
+                {"rejection_reason": body_params.rejection_reason}
+            ),
         )
     logger.error("Email action type %s is not implemented", action_type)
     raise EmailDispatchException(f"Email action type {action_type} is not implemented")